+++ src/main/java/com/programmers/heycake/domain/member/controller/MemberController.java
@@ -1,7 +1,5 @@
 package com.programmers.heycake.domain.member.controller;
 
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
 import org.springframework.http.ResponseEntity;
@@ -9,10 +7,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.programmers.heycake.domain.facade.MemberFacade;
 import com.programmers.heycake.domain.member.model.dto.request.AuthenticationCodeRequest;
 import com.programmers.heycake.domain.member.model.dto.request.TokenRefreshRequest;
 import com.programmers.heycake.domain.member.model.dto.response.TokenResponse;
-import com.programmers.heycake.domain.member.service.MemberService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,33 +19,26 @@ import lombok.extern.slf4j.Slf4j;
 @RestController
 @RequiredArgsConstructor
 public class MemberController {
-
-	private final MemberService memberService;
+	private final MemberFacade memberFacade;
 
 	@PostMapping("/login")
 	public ResponseEntity<TokenResponse> login(
 			@RequestBody AuthenticationCodeRequest authenticationCodeRequest) {
-		TokenResponse tokenResponse = memberService.loginForKakao(authenticationCodeRequest.code());
-
+		TokenResponse tokenResponse = memberFacade.loginForKakao(authenticationCodeRequest.code());
 		return ResponseEntity.ok(tokenResponse);
 	}
 
 	@PostMapping("/logout")
 	public ResponseEntity<Void> logout() {
-		memberService.logout();
+		memberFacade.logout();
 		return ResponseEntity.noContent().build();
 	}
 
 	@PostMapping("api/v1/members/refresh")
 	public ResponseEntity<String> refreshToken(
-			@Valid @RequestBody TokenRefreshRequest tokenRefreshRequest,
-			HttpServletResponse response
+			@Valid @RequestBody TokenRefreshRequest tokenRefreshRequest
 	) {
-		TokenResponse tokenResponse = memberService.reissueToken(tokenRefreshRequest.refreshToken());
-		Cookie accessToken = new Cookie("access_token", tokenResponse.accessToken());
-		accessToken.setPath("/");
-		accessToken.setHttpOnly(true);
-		response.addCookie(accessToken);
+		TokenResponse tokenResponse = memberFacade.reissueToken(tokenRefreshRequest.refreshToken());
 		return ResponseEntity.ok(tokenResponse.refreshToken());
 	}
 }
